requirements.txt
bert-score==0.3.13
bitsandbytes==0.46.0
chardet==5.2.0
charset-normalizer==3.3.2
evaluate==0.4.3
executing==0.8.3
google-auth-oauthlib==0.4.1
hf-xet==1.1.4
nltk==3.9.1
numexpr==2.10.1
numpy==2.2.5
optuna==4.4.0
peft==0.15.2
plotly==6.1.2
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
torch==2.7.1+cu128